Move all the config into the TypelevelTheme
